выполнена модуляция кода
